@@ -1,16 +1,16 @@
 package lambroll
 
 import (
+	"context"
+	"fmt"
 	"log"
 
-	"github.com/aws/aws-sdk-go/service/lambda"
-	"github.com/pkg/errors"
+	"github.com/aws/aws-sdk-go-v2/service/lambda"
 )
 
 // DeleteOption represents options for Delete()
 type DeleteOption struct {
-	FunctionFilePath *string
-	DryRun           *bool
+	DryRun *bool
 }
 
 func (opt DeleteOption) label() string {
@@ -21,10 +21,10 @@ func (opt DeleteOption) label() string {
 }
 
 // Delete deletes function
-func (app *App) Delete(opt DeleteOption) error {
-	fn, err := app.loadFunction(*opt.FunctionFilePath)
+func (app *App) Delete(ctx context.Context, opt DeleteOption) error {
+	fn, err := app.loadFunction(app.functionFilePath)
 	if err != nil {
-		return errors.Wrap(err, "failed to load function")
+		return fmt.Errorf("failed to load function: %w", err)
 	}
 
 	log.Println("[info] deleting function", *fn.FunctionName, opt.label())
@@ -32,11 +32,11 @@ func (app *App) Delete(opt DeleteOption) error {
 	if *opt.DryRun {
 		return nil
 	}
-	_, err = app.lambda.DeleteFunction(&lambda.DeleteFunctionInput{
+	_, err = app.lambda.DeleteFunction(ctx, &lambda.DeleteFunctionInput{
 		FunctionName: fn.FunctionName,
 	})
 	if err != nil {
-		return errors.Wrap(err, "failed to delete function")
+		return fmt.Errorf("failed to delete function: %w", err)
 	}
 
 	return nil
