@@ -1,6 +1,8 @@
 package lambroll
 
 import (
+	"context"
+	"errors"
 	"fmt"
 	"io"
 	"log"
@@ -8,49 +10,46 @@ import (
 	"os"
 	"strings"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
-	"github.com/aws/aws-sdk-go/service/lambda"
-	"github.com/pkg/errors"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/lambda"
+	"github.com/aws/aws-sdk-go-v2/service/lambda/types"
 )
 
 // InitOption represents options for Init()
 type InitOption struct {
-	FunctionName *string
-	DownloadZip  *bool
+	FunctionName *string `help:"Function name for init" required:"true" default:""`
+	DownloadZip  bool    `help:"Download function.zip" default:"false"`
+	Jsonnet      bool    `default:"false" help:"render function.json as jsonnet"`
 }
 
 // Init initializes function.json
-func (app *App) Init(opt InitOption) error {
-	res, err := app.lambda.GetFunction(&lambda.GetFunctionInput{
+func (app *App) Init(ctx context.Context, opt *InitOption) error {
+	res, err := app.lambda.GetFunction(ctx, &lambda.GetFunctionInput{
 		FunctionName: opt.FunctionName,
 	})
-	var c *lambda.FunctionConfiguration
+	var c *types.FunctionConfiguration
 	exists := true
 	if err != nil {
-		if aerr, ok := err.(awserr.Error); ok {
-			switch aerr.Code() {
-			case lambda.ErrCodeResourceNotFoundException:
-				log.Printf("[info] function %s is not found", *opt.FunctionName)
-				c = &lambda.FunctionConfiguration{
-					FunctionName: opt.FunctionName,
-					MemorySize:   aws.Int64(128),
-					Runtime:      aws.String("nodejs18.x"),
-					Timeout:      aws.Int64(3),
-					Handler:      aws.String("index.handler"),
-					Role: aws.String(
-						fmt.Sprintf(
-							"arn:aws:iam::%s:role/YOUR_LAMBDA_ROLE_NAME",
-							app.AWSAccountID(),
-						),
+		var nfe *types.ResourceNotFoundException
+		if errors.As(err, &nfe) {
+			log.Printf("[info] function %s is not found", *opt.FunctionName)
+			c = &types.FunctionConfiguration{
+				FunctionName: opt.FunctionName,
+				MemorySize:   aws.Int32(128),
+				Runtime:      types.RuntimeNodejs18x,
+				Timeout:      aws.Int32(3),
+				Handler:      aws.String("index.handler"),
+				Role: aws.String(
+					fmt.Sprintf(
+						"arn:aws:iam::%s:role/YOUR_LAMBDA_ROLE_NAME",
+						app.AWSAccountID(ctx),
 					),
-				}
-				exists = false
-			default:
+				),
 			}
+			exists = false
 		}
 		if c == nil {
-			return errors.Wrap(err, "failed to GetFunction"+*opt.FunctionName)
+			return fmt.Errorf("failed to GetFunction %s: %w", *opt.FunctionName, err)
 		}
 	} else {
 		log.Printf("[info] function %s found", *opt.FunctionName)
@@ -59,20 +58,20 @@ func (app *App) Init(opt InitOption) error {
 
 	var tags Tags
 	if exists {
-		arn := app.functionArn(*c.FunctionName)
+		arn := app.functionArn(ctx, *c.FunctionName)
 		log.Printf("[debug] listing tags of %s", arn)
-		res, err := app.lambda.ListTags(&lambda.ListTagsInput{
+		res, err := app.lambda.ListTags(ctx, &lambda.ListTagsInput{
 			Resource: aws.String(arn),
 		})
 		if err != nil {
-			return errors.Wrap(err, "faled to list tags")
+			return fmt.Errorf("faled to list tags: %w", err)
 		}
 		tags = res.Tags
 	}
 
 	fn := newFunctionFrom(c, res.Code, tags)
 
-	if *opt.DownloadZip && res.Code != nil && *res.Code.RepositoryType == "S3" {
+	if opt.DownloadZip && res.Code != nil && *res.Code.RepositoryType == "S3" {
 		log.Printf("[info] downloading %s", FunctionZipFilename)
 		if err := download(*res.Code.Location, FunctionZipFilename); err != nil {
 			return err
@@ -89,20 +88,32 @@ func (app *App) Init(opt InitOption) error {
 		return err
 	}
 
-	log.Printf("[info] creating %s", FunctionFilenames[0])
+	var name string
+	if opt.Jsonnet {
+		name = DefaultFunctionFilenames[1]
+	} else {
+		name = DefaultFunctionFilenames[0]
+	}
+	log.Printf("[info] creating %s", name)
 	b, _ := marshalJSON(fn)
-	return app.saveFile(FunctionFilenames[0], b, os.FileMode(0644))
+	if opt.Jsonnet {
+		b, err = jsonToJsonnet(b, name)
+		if err != nil {
+			return err
+		}
+	}
+	return app.saveFile(name, b, os.FileMode(0644))
 }
 
 func download(url, path string) error {
 	resp, err := http.Get(url)
 	if err != nil {
-		return errors.Wrapf(err, "failed to get %s", url)
+		return fmt.Errorf("failed to get %s: %w", url, err)
 	}
 	defer resp.Body.Close()
 	f, err := os.OpenFile(path, os.O_RDWR|os.O_CREATE, os.FileMode(0644))
 	if err != nil {
-		return errors.Wrapf(err, "failed to open file %s", path)
+		return fmt.Errorf("failed to open file %s: %w", path, err)
 	}
 	_, err = io.Copy(f, resp.Body)
 	return err
